Add transparancy control to pdos (#33)
The pdos lines plot doesn't have options to set the background color. 
In this PR the setting is added.
The default is set to transparent to only draw the lines.
The parameter setting for this field is strict to the string of a percentage value between 0% to 100%, where 0% for fully transparent.